refactor(mir): architecture pass — split lowering and validation into module directories
The pass the crate-level lint suppressions promised:

- lower.rs (3,787 lines) becomes lower/{mod,context,expr,stmt,place,generics}.rs
  with lower_hir decomposed into phase helpers and the 9/12-argument
  context constructors replaced by LoweringShared + FunctionSpec structs.
- validate.rs (1,939 lines) becomes validate/{mod,structure,operands,
  assignment,closures}.rs with validate_function and the definite-
  assignment dataflow split into phase helpers.
- format.rs assembles output through a fmt::Write helper instead of
  push_str(&format!(..)).
- Shadowing renames, map_or/is_ok_and rewrites, const fns, and an
  # Errors section for lower_hir across the crate.
- lib.rs suppressions drop from 13 to 5; the survivors (too_many_lines,
  match_same_arms, similar_names, exhaustive_enums/structs) now carry
  reasons describing the deliberate design they cover: exhaustive
  ExprKind/Rvalue walkers stay single matches so variant additions are
  compile-time review sites.

cargo clippy -p smelt-mir: zero warnings. Tests green:
smelt-mir 10, smelt-frontend-ts 674, smelt-specialize 28, smelt-hir 3.
