+++ crates/smelt-mir/src/validate/operands.rs
@@ -1,205 +1,15 @@
-use std::collections::{HashSet, VecDeque};
+//! Operand traversal and existence checks.
+//!
+//! This module owns the canonical operand walk for [`Rvalue`]
+//! ([`Rvalue::for_each_operand`] and its mutable mirror), which the rest of the
+//! crate reuses so operand enumeration cannot drift as variants are added. It
+//! also holds the "does this reference exist" checks for rvalues, operands,
+//! places, and callees used by structural validation.
 
-use smelt_hir::TypeId;
+use crate::{Callee, MirFunction, Mir, Operand, Place, Rvalue};
 
-use crate::{Callee, LocalId, Mir, MirFunction, Operand, Place, Rvalue, Statement, Terminator};
-
-/// A validation error discovered while checking MIR.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ValidationError {
-    /// Human-readable description of the problem.
-    pub message: String,
-}
-
-#[must_use]
-/// Validates MIR and returns any structural errors.
-pub fn validate(mir: &Mir) -> Vec<ValidationError> {
-    let mut errors = Vec::new();
-    validate_closures(mir, &mut errors);
-    for function in &mir.functions {
-        validate_function(mir, function, &mut errors);
-    }
-    errors
-}
-
-/// Validate MIR closure table entries and escape-specific capture rules.
-///
-/// The escaping-closure capture check below enforces invariant 1 of the closure
-/// ABI contract published by the `lower::closures` module: every closure with
-/// `escapes == true` must capture by value so Rust codegen can emit an owning
-/// `move` closure that does not borrow stack locals past their owner's return.
-fn validate_closures(mir: &Mir, errors: &mut Vec<ValidationError>) {
-    for (idx, closure) in mir.closures.iter().enumerate() {
-        if usize::try_from(closure.id.0).ok() != Some(idx) {
-            errors.push(error(format!(
-                "closure index {idx} has mismatched id {:?}",
-                closure.id
-            )));
-        }
-        for local in &closure.locals {
-            validate_type(mir, local.ty, errors);
-        }
-        for param in &closure.params {
-            if usize::try_from(param.0)
-                .ok()
-                .and_then(|index| closure.locals.get(index))
-                .is_none()
-            {
-                errors.push(error(format!(
-                    "closure {:?} parameter references unknown local {:?}",
-                    closure.id, param
-                )));
-            }
-        }
-        for capture in &closure.captures {
-            validate_type(mir, capture.ty, errors);
-            if let Some(target) = capture.target_local
-                && usize::try_from(target.0)
-                    .ok()
-                    .and_then(|index| closure.locals.get(index))
-                    .is_none()
-            {
-                errors.push(error(format!(
-                    "closure {:?} capture targets unknown local {:?}",
-                    closure.id, target
-                )));
-            }
-            // Closure ABI contract invariant 1 (see `lower::closures`):
-            // escaping closures capture by value only.
-            if closure.escapes && capture.mode != smelt_hir::CaptureMode::ByValue {
-                errors.push(error(format!(
-                    "escaping closure {:?} captures {:?} without owning it",
-                    closure.id, capture.source_local
-                )));
-            }
-        }
-        validate_type(mir, closure.return_ty, errors);
-    }
-}
-
-/// Validate one MIR function and append any discovered errors.
-fn validate_function(mir: &Mir, function: &MirFunction, errors: &mut Vec<ValidationError>) {
-    let entry_idx = block_index(function.entry);
-    if function.blocks.get(entry_idx).is_none() {
-        errors.push(error(format!(
-            "function {:?} has an unknown entry block {:?}",
-            function.id, function.entry
-        )));
-    }
-
-    for (block_idx, block) in function.blocks.iter().enumerate() {
-        if block_index(block.id) != block_idx {
-            errors.push(error(format!(
-                "function {:?} block index {block_idx} has mismatched id {:?}",
-                function.id, block.id
-            )));
-        }
-        if block.terminator.is_none() {
-            errors.push(error(format!(
-                "function {:?} block {:?} is missing a terminator",
-                function.id, block.id
-            )));
-        }
-
-        for phi in &block.phis {
-            validate_type(mir, phi.ty, errors);
-            validate_local_exists(function, phi.dest, errors);
-            for (_, operand) in &phi.incoming {
-                validate_operand_exists(function, operand, errors);
-            }
-        }
-
-        for stmt in &block.statements {
-            match stmt {
-                Statement::Assign { dest, value } => {
-                    validate_rvalue_exists(mir, function, value, errors);
-                    validate_local_exists(function, *dest, errors);
-                }
-                Statement::AssignPlace { place, value } => {
-                    validate_place_exists(function, place, errors);
-                    validate_rvalue_exists(mir, function, value, errors);
-                }
-                Statement::StorageLive(local) | Statement::StorageDead(local) => {
-                    validate_local_exists(function, *local, errors);
-                }
-            }
-        }
-
-        if let Some(terminator) = &block.terminator {
-            match terminator {
-                Terminator::Goto(target) => validate_block_exists(function, *target, errors),
-                Terminator::Call {
-                    callee,
-                    args,
-                    dest,
-                    target,
-                    unwind,
-                } => {
-                    validate_callee_exists(mir, function, callee, errors);
-                    for arg in args {
-                        validate_operand_exists(function, arg, errors);
-                    }
-                    validate_local_exists(function, *dest, errors);
-                    validate_block_exists(function, *target, errors);
-                    if let Some(handler) = unwind {
-                        validate_block_exists(function, handler.catch_block, errors);
-                        if let Some(local) = handler.exception_local {
-                            validate_local_exists(function, local, errors);
-                        }
-                    }
-                }
-                Terminator::Await {
-                    future,
-                    dest,
-                    target,
-                    unwind,
-                } => {
-                    validate_operand_exists(function, future, errors);
-                    validate_local_exists(function, *dest, errors);
-                    validate_block_exists(function, *target, errors);
-                    if let Some(handler) = unwind {
-                        validate_block_exists(function, handler.catch_block, errors);
-                        if let Some(local) = handler.exception_local {
-                            validate_local_exists(function, local, errors);
-                        }
-                    }
-                }
-                Terminator::Switch {
-                    cond,
-                    then_block,
-                    else_block,
-                } => {
-                    validate_operand_exists(function, cond, errors);
-                    validate_block_exists(function, *then_block, errors);
-                    validate_block_exists(function, *else_block, errors);
-                }
-                Terminator::Match {
-                    scrutinee,
-                    arms,
-                    default,
-                } => {
-                    validate_operand_exists(function, scrutinee, errors);
-                    for arm in arms {
-                        validate_block_exists(function, arm.target, errors);
-                    }
-                    if let Some(default_target) = default {
-                        validate_block_exists(function, *default_target, errors);
-                    }
-                }
-                Terminator::Return(operand) | Terminator::Throw(operand) => {
-                    validate_operand_exists(function, operand, errors);
-                }
-                Terminator::Unreachable => {}
-            }
-        }
-    }
-
-    for local in &function.locals {
-        validate_type(mir, local.ty, errors);
-    }
-
-    validate_definite_assignment(mir, function, errors);
-}
+use super::structure::validate_local_exists;
+use super::{ValidationError, error, function_index, validate_type};
 
 impl Rvalue {
     /// Visit every operand read by this rvalue in evaluation order.
@@ -363,11 +173,11 @@ impl Rvalue {
                 end,
             } => {
                 visit(operand);
-                if let Some(operand) = start.as_ref() {
-                    visit(operand);
+                if let Some(start_operand) = start.as_ref() {
+                    visit(start_operand);
                 }
-                if let Some(operand) = end.as_ref() {
-                    visit(operand);
+                if let Some(end_operand) = end.as_ref() {
+                    visit(end_operand);
                 }
             }
             Self::ListContains { list, item } => {
@@ -1022,11 +832,11 @@ impl Rvalue {
                 end,
             } => {
                 visit(operand);
-                if let Some(operand) = start.as_mut() {
-                    visit(operand);
+                if let Some(start_operand) = start.as_mut() {
+                    visit(start_operand);
                 }
-                if let Some(operand) = end.as_mut() {
-                    visit(operand);
+                if let Some(end_operand) = end.as_mut() {
+                    visit(end_operand);
                 }
             }
             Self::ListContains { list, item } => {
@@ -1519,7 +1329,7 @@ impl Rvalue {
 }
 
 /// Validate that IDs referenced by an rvalue point to existing MIR entities.
-fn validate_rvalue_exists(
+pub(super) fn validate_rvalue_exists(
     mir: &Mir,
     function: &MirFunction,
     value: &Rvalue,
@@ -1532,9 +1342,9 @@ fn validate_rvalue_exists(
             .get(usize::try_from(id.0).unwrap_or(usize::MAX))
             .is_none()
     {
-        errors.push(ValidationError {
-            message: format!("closure rvalue references unknown closure {id:?}"),
-        });
+        errors.push(error(format!(
+            "closure rvalue references unknown closure {id:?}"
+        )));
     }
     if let Rvalue::UnknownCast { target, .. } = value {
         validate_type(mir, *target, errors);
@@ -1542,7 +1352,7 @@ fn validate_rvalue_exists(
 }
 
 /// Validate that IDs referenced by an operand point to existing MIR entities.
-fn validate_operand_exists(
+pub(super) fn validate_operand_exists(
     function: &MirFunction,
     operand: &Operand,
     errors: &mut Vec<ValidationError>,
@@ -1556,7 +1366,11 @@ fn validate_operand_exists(
 }
 
 /// Validate that a place references valid locals and projected fields.
-fn validate_place_exists(function: &MirFunction, place: &Place, errors: &mut Vec<ValidationError>) {
+pub(super) fn validate_place_exists(
+    function: &MirFunction,
+    place: &Place,
+    errors: &mut Vec<ValidationError>,
+) {
     match place {
         Place::Local(local) => {
             validate_local_exists(function, *local, errors);
@@ -1570,7 +1384,7 @@ fn validate_place_exists(function: &MirFunction, place: &Place, errors: &mut Vec
 }
 
 /// Validate that a callee target exists and references valid receiver places.
-fn validate_callee_exists(
+pub(super) fn validate_callee_exists(
     mir: &Mir,
     function: &MirFunction,
     callee: &Callee,
@@ -1587,353 +1401,4 @@ fn validate_callee_exists(
         }
         Callee::Builtin(_) => {}
     }
-}
-
-/// Perform forward dataflow to ensure locals are assigned before use.
-fn validate_definite_assignment(
-    mir: &Mir,
-    function: &MirFunction,
-    errors: &mut Vec<ValidationError>,
-) {
-    let entry_idx = block_index(function.entry);
-    if function.blocks.get(entry_idx).is_none() {
-        return;
-    }
-
-    let block_count = function.blocks.len();
-    let mut in_sets = vec![None::<HashSet<LocalId>>; block_count];
-    let mut queue = VecDeque::new();
-    let mut entry_defs = HashSet::new();
-    entry_defs.extend(function.params.iter().copied());
-    entry_defs.extend(function.locals.iter().enumerate().filter_map(
-        |(idx, local)| match local.kind {
-            crate::LocalKind::UserBinding(_) => Some(LocalId(u32::try_from(idx).ok()?)),
-            crate::LocalKind::Param { .. } | crate::LocalKind::Temp => None,
-        },
-    ));
-    if let Some(slot) = in_sets.get_mut(entry_idx) {
-        *slot = Some(entry_defs);
-    } else {
-        return;
-    }
-    queue.push_back(function.entry);
-
-    while let Some(block_id) = queue.pop_front() {
-        let block_idx = block_index(block_id);
-        let Some(block) = function.blocks.get(block_idx) else {
-            continue;
-        };
-        let Some(definitions_slot) = in_sets.get(block_idx) else {
-            continue;
-        };
-        let Some(mut definitions) = definitions_slot.clone() else {
-            continue;
-        };
-
-        for phi in &block.phis {
-            definitions.insert(phi.dest);
-        }
-
-        for stmt in &block.statements {
-            match stmt {
-                Statement::Assign { dest, value } => {
-                    let _ = value;
-                    definitions.insert(*dest);
-                }
-                Statement::AssignPlace { place, value } => {
-                    let _ = value;
-                    match place {
-                        Place::Local(local) => {
-                            definitions.insert(*local);
-                        }
-                        Place::Field { .. } | Place::Index { .. } => {}
-                    }
-                }
-                Statement::StorageLive(_) | Statement::StorageDead(_) => {}
-            }
-        }
-
-        if let Some(terminator) = &block.terminator {
-            match terminator {
-                Terminator::Goto(_) | Terminator::Unreachable => {}
-                Terminator::Call { dest, .. } | Terminator::Await { dest, .. } => {
-                    definitions.insert(*dest);
-                }
-                Terminator::Switch { .. }
-                | Terminator::Match { .. }
-                | Terminator::Return(_)
-                | Terminator::Throw(_) => {}
-            }
-
-            for successor in terminator.successors() {
-                let Some(existing) = in_sets.get_mut(block_index(successor)) else {
-                    continue;
-                };
-                let changed = if let Some(existing_defs) = existing {
-                    let intersection = existing_defs
-                        .intersection(&definitions)
-                        .copied()
-                        .collect::<HashSet<_>>();
-                    let changed = *existing_defs != intersection;
-                    *existing_defs = intersection;
-                    changed
-                } else {
-                    *existing = Some(definitions.clone());
-                    true
-                };
-                if changed {
-                    queue.push_back(successor);
-                }
-            }
-        }
-    }
-
-    for (block_idx, block) in function.blocks.iter().enumerate() {
-        let Some(definitions_slot) = in_sets.get(block_idx) else {
-            continue;
-        };
-        let Some(mut definitions) = definitions_slot.clone() else {
-            continue;
-        };
-        let before_phi = definitions.clone();
-
-        for phi in &block.phis {
-            for (_, operand) in &phi.incoming {
-                validate_operand(mir, function, &before_phi, operand, errors);
-            }
-            definitions.insert(phi.dest);
-        }
-
-        for stmt in &block.statements {
-            match stmt {
-                Statement::Assign { dest, value } => {
-                    validate_rvalue(mir, function, &definitions, value, errors);
-                    definitions.insert(*dest);
-                }
-                Statement::AssignPlace { place, value } => {
-                    validate_rvalue(mir, function, &definitions, value, errors);
-                    match place {
-                        Place::Local(local) => {
-                            definitions.insert(*local);
-                        }
-                        Place::Field { .. } | Place::Index { .. } => {
-                            validate_place(mir, function, &definitions, place, errors);
-                        }
-                    }
-                }
-                Statement::StorageLive(_) | Statement::StorageDead(_) => {}
-            }
-        }
-
-        if let Some(terminator) = &block.terminator {
-            match terminator {
-                Terminator::Goto(_) | Terminator::Unreachable => {}
-                Terminator::Call {
-                    callee,
-                    args,
-                    dest,
-                    unwind,
-                    ..
-                } => {
-                    validate_callee(mir, function, &definitions, callee, errors);
-                    for arg in args {
-                        validate_operand(mir, function, &definitions, arg, errors);
-                    }
-                    if let Some(handler) = unwind
-                        && let Some(local) = handler.exception_local
-                    {
-                        validate_local_exists(function, local, errors);
-                    }
-                    definitions.insert(*dest);
-                }
-                Terminator::Await {
-                    future,
-                    dest,
-                    unwind,
-                    ..
-                } => {
-                    validate_operand(mir, function, &definitions, future, errors);
-                    if let Some(handler) = unwind
-                        && let Some(local) = handler.exception_local
-                    {
-                        validate_local_exists(function, local, errors);
-                    }
-                    definitions.insert(*dest);
-                }
-                Terminator::Switch { cond, .. } => {
-                    validate_operand(mir, function, &definitions, cond, errors);
-                }
-                Terminator::Match { scrutinee, .. } => {
-                    validate_operand(mir, function, &definitions, scrutinee, errors);
-                }
-                Terminator::Return(operand) | Terminator::Throw(operand) => {
-                    validate_operand(mir, function, &definitions, operand, errors);
-                }
-            }
-        }
-    }
-}
-
-/// Validate type constraints for one rvalue.
-fn validate_rvalue(
-    mir: &Mir,
-    function: &MirFunction,
-    definitions: &HashSet<LocalId>,
-    value: &Rvalue,
-    errors: &mut Vec<ValidationError>,
-) {
-    // Operand traversal is the single source of truth in `Rvalue::for_each_operand`;
-    // route validation through it so the two cannot drift as variants are added.
-    value.for_each_operand(|operand| {
-        validate_operand(mir, function, definitions, operand, errors);
-    });
-}
-
-/// Validate type constraints for one operand.
-fn validate_operand(
-    mir: &Mir,
-    function: &MirFunction,
-    definitions: &HashSet<LocalId>,
-    operand: &Operand,
-    errors: &mut Vec<ValidationError>,
-) {
-    match operand {
-        Operand::Copy(place) | Operand::Move(place) => {
-            validate_place(mir, function, definitions, place, errors);
-        }
-        Operand::Const(_) => {}
-    }
-}
-
-/// Validate type constraints for one place projection chain.
-fn validate_place(
-    mir: &Mir,
-    function: &MirFunction,
-    definitions: &HashSet<LocalId>,
-    place: &Place,
-    errors: &mut Vec<ValidationError>,
-) {
-    match place {
-        Place::Local(local) => {
-            if !definitions.contains(local) {
-                let function_name = mir.symbols.get(function.name).unwrap_or("<unknown>");
-                let local_detail = local_detail(mir, function, *local);
-                errors.push(error(format!(
-                    "function {:?} `{function_name}` reads local {:?} {local_detail} before it is definitely defined",
-                    function.id, local
-                )));
-            }
-        }
-        Place::Field { base, .. } => {
-            validate_place(mir, function, definitions, &Place::Local(*base), errors)
-        }
-        Place::Index { base, index } => {
-            validate_place(mir, function, definitions, &Place::Local(*base), errors);
-            validate_operand(mir, function, definitions, index, errors);
-        }
-    }
-}
-
-/// Render a local declaration for definite-assignment diagnostics.
-fn local_detail(mir: &Mir, function: &MirFunction, local: LocalId) -> String {
-    let Some(decl) = function.locals.get(local_index(local)) else {
-        return "<unknown local>".to_owned();
-    };
-    let kind = match decl.kind {
-        crate::LocalKind::Param { symbol } => symbol
-            .and_then(|param_symbol| mir.symbols.get(param_symbol))
-            .map_or_else(|| "param".to_owned(), |name| format!("param `{name}`")),
-        crate::LocalKind::Temp => "temp".to_owned(),
-        crate::LocalKind::UserBinding(symbol) => {
-            format!(
-                "binding `{}`",
-                mir.symbols.get(symbol).unwrap_or("<unknown>")
-            )
-        }
-    };
-    format!(
-        "({kind}, span file {} start {} end {})",
-        decl.span.file.0, decl.span.start, decl.span.end
-    )
-}
-
-/// Validate type constraints for one callee.
-fn validate_callee(
-    mir: &Mir,
-    function: &MirFunction,
-    definitions: &HashSet<LocalId>,
-    callee: &Callee,
-    errors: &mut Vec<ValidationError>,
-) {
-    match callee {
-        Callee::Static(_) | Callee::Builtin(_) => {}
-        Callee::Indirect(operand) => {
-            validate_operand(mir, function, definitions, operand, errors);
-        }
-    }
-}
-
-/// Ensure a local ID points to an existing local declaration.
-fn validate_local_exists(
-    function: &MirFunction,
-    local: LocalId,
-    errors: &mut Vec<ValidationError>,
-) {
-    if function.locals.get(local_index(local)).is_none() {
-        errors.push(error(format!(
-            "function {:?} references unknown local {:?}",
-            function.id, local
-        )));
-    }
-}
-
-/// Ensure a block ID points to an existing basic block.
-fn validate_block_exists(
-    function: &MirFunction,
-    block: crate::BlockId,
-    errors: &mut Vec<ValidationError>,
-) {
-    if function.blocks.get(block_index(block)).is_none() {
-        errors.push(error(format!(
-            "function {:?} references unknown block {:?}",
-            function.id, block
-        )));
-    }
-}
-
-/// Ensure a type ID points to an existing MIR type entry.
-fn validate_type(mir: &Mir, ty: TypeId, errors: &mut Vec<ValidationError>) {
-    if mir.types.get(ty).is_none() {
-        errors.push(error(format!("MIR references unknown type {ty:?}")));
-    }
-}
-
-/// Convert a local identifier into a vector index.
-fn local_index(local: LocalId) -> usize {
-    u32_to_usize(local.0, "MIR local id")
-}
-
-/// Convert a block identifier into a vector index.
-fn block_index(block: crate::BlockId) -> usize {
-    u32_to_usize(block.0, "MIR block id")
-}
-
-/// Convert a function identifier into a vector index.
-fn function_index(function: crate::FuncId) -> usize {
-    u32_to_usize(function.0, "MIR function id")
-}
-
-/// Convert a `u32` identifier into a vector index.
-fn u32_to_usize(value: u32, label: &str) -> usize {
-    if let Ok(index) = usize::try_from(value) {
-        index
-    } else {
-        let _ = label;
-        usize::MAX
-    }
-}
-
-/// Build a validation error with no source span.
-fn error(message: String) -> ValidationError {
-    ValidationError { message }
 }
